Updated code to 640 final release

+++ source/developr.c
@@ -1,7 +1,7 @@
    /*******************************************************/
    /*      "C" Language Integrated Production System      */
    /*                                                     */
-   /*            CLIPS Version 6.40  11/01/16             */
+   /*            CLIPS Version 6.40  04/08/20             */
    /*                                                     */
    /*                   DEVELOPER MODULE                  */
    /*******************************************************/
@@ -493,7 +493,7 @@ static void PrintOPNLevel(
   {
    CLASS_BITMAP *cbmp;
    SLOT_BITMAP *sbmp;
-   unsigned i;
+   unsigned short i;
    OBJECT_PATTERN_NODE *uptr;
    OBJECT_ALPHA_NODE *alphaPtr;
 
@@ -576,7 +576,7 @@ void InstanceTableUsageCommand(
   UDFValue *returnValue)
   {
    unsigned long i;
-   int instanceCounts[COUNT_SIZE];
+   unsigned long long instanceCounts[COUNT_SIZE];
    Instance *ins;
    unsigned long instanceCount, totalInstanceCount = 0;
